Rename 'projectUrl' folder to 'selectProjectUrl'
